Add the NDK analyzer scripts.
These are used by the NDK for scan-build so it is possible to do an
analysis run on a GCC project. The problem is that it can't use clang
directly with GCC's cflags because GCC doesn't use the -target flag,
and clang can't cross compile without it.

Longer term the analyzer handling should probably be pushed down into
ndk-build, and when that happens these can be removed.

Bug: http://b/22801531
Change-Id: Ifd869520a8331adaade925e0077671f32bb8ff1f